+++ Tara/Models/ChatModels/FirestoreModel.swift
@@ -171,7 +171,7 @@ struct Message{
     public var tempImage : UIImage = UIImage()
     public var thumbnailURL : String = ""
     public var messageDuration : Int = 0
-    public var price : Int = 0
+    public var price : Double = 0
     
     //MARK:- Inits
     //=============
@@ -191,7 +191,7 @@ struct Message{
         self.thumbnailURL = dict[ApiKey.thumbnail] as? String ?? ""
         self.blocked = dict[ApiKey.blocked] as? Bool ?? false
         self.messageDuration =  dict[ApiKey.messageDuration] as? Int ?? 0
-        self.price = dict[ApiKey.price] as? Int ?? 0
+        self.price = dict[ApiKey.price] as? Double ?? 0.0
       }
 }
